+++ src/models/llama.py
@@ -219,7 +219,7 @@ class Llama(GPTBase):
             # if we are given some desired targets also calculate the loss
             logits = self.lm_head(x)
             loss = F.cross_entropy(
-                logits.view(-1, logits.size(-1)), targets.view(-1), ignore_index=-1
+                logits.view(-1, logits.size(-1)), targets.view(-1).to(torch.long), ignore_index=-1
             )
         else:
             # inference-time mini-optimization: only forward the lm_head on the very last position
